fix:set tag group as optional
